--- nilearn/experimental/surface/__init__.py
@@ -1,31 +1,0 @@
-"""The :mod:`nilearn.experimental.surface` module."""
-
-from nilearn.experimental.surface._datasets import (
-    fetch_destrieux,
-    fetch_nki,
-    load_fsaverage,
-)
-from nilearn.experimental.surface._maskers import (
-    SurfaceLabelsMasker,
-    SurfaceMasker,
-)
-from nilearn.experimental.surface._surface_image import (
-    FileMesh,
-    InMemoryMesh,
-    Mesh,
-    PolyMesh,
-    SurfaceImage,
-)
-
-__all__ = [
-    "FileMesh",
-    "InMemoryMesh",
-    "Mesh",
-    "PolyMesh",
-    "SurfaceImage",
-    "SurfaceLabelsMasker",
-    "SurfaceMasker",
-    "fetch_destrieux",
-    "fetch_nki",
-    "load_fsaverage",
-]
